+++ src/main/webapp/resources/css/modal.css
@@ -99,8 +99,8 @@
     font-weight: 100;
     width:270px;
     
-    right: 10px;
-    bottom: 0px;
+    right: 15px;
+    bottom: 20px;
     
     padding: 15px 20px;
     background-color: #EBEBEB;
